Move things out of senpai
* Not having a sub-module helps resolve the idiotic package import system that is PYTHONPATH :disappointed:
